# optimize IsExpressionTree