ImmutabilityAnalyzer for reference

## Before
```
D2L.CodeStyle.Analyzers.Immutability.ImmutabilityAnalyzer = 11.804 s
D2L.CodeStyle.Analyzers.Language.RequireNamedArgumentsAnalyzer = 9.463 s
```

##After
```
D2L.CodeStyle.Analyzers.Immutability.ImmutabilityAnalyzer = 14.822 s
D2L.CodeStyle.Analyzers.Language.RequireNamedArgumentsAnalyzer = 4.902 s
```